Taxsim

Taxsim.taxsim32

Taxsim.taxsim32 — Method

Before using taxsim32, please make yourself familiar with Internet TAXSIM 32. Submit a few individual observations and upload an entire csv file.

Syntax

taxsim32(df; kwargs...)

df has to be a DataFrame object with at least one observation.

Included columns have to be named exactly as in the Internet TAXSIM 32 variable list (bold names after boxes) but can be in any order. taxsim32 returns typos and case errors.

Non-provided input variables are set to zero by the TAXSIM server but " " (blanks as strings) or missing lead to non-response as the server only accepts Integers or Floats. taxsim32 returns type errors.

Keyword Arguments

connection: choose either "FTP" or "SSH". Defaults to "FTP".

full: request the full list of TAXSIM return variables v1 to v42. Defaults to false which returns v1 to v9.

long_names: name all return variables with their long TAXSIM names (as opposed to abbreviated names for v1 to v9 and no names for v10 to v42). Defaults to false.

Output

Data frame with requested TAXSIM return variables. Column types are either Integer or Float.

If df does not include state or if state = 0, then the returned data frame does not include v30 to v42.

Observations are ordered as in df so hcat(df, df_output, makeunique=true) merges all variables of the input and output data frames.

Examples

using DataFrames, Taxsim

df_input_small = DataFrame(year=1980, state=5, mstat=2, ltcg=100000)
1×4 DataFrame
│ Row │ year  │ state │ mstat │ ltcg   │
│     │ Int64 │ Int64 │ Int64 │ Int64  │
├─────┼───────┼───────┼───────┼────────┤
│ 1   │ 1980  │ 5     │ 2     │ 100000 │

df_output1 = taxsim32(df_input_small)
1×9 DataFrame
│ Row │ taxsimid │ year  │ state │ fiitax  │ siitax  │ fica    │ frate   │ srate   │ ficar   │
│     │ Float64  │ Int64 │ Int64 │ Float64 │ Float64 │ Float64 │ Float64 │ Float64 │ Float64 │
├─────┼──────────┼───────┼───────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ 1   │ 0.0      │ 1980  │ 5     │ 10701.2 │ 4493.8  │ 0.0     │ 17.8    │ 11.0    │ 12.0    │

df_output2 = taxsim32(df_input_small, connection="SSH", full=true, long_lables=true);